PLANT-DISEASE-IDENTIFICATION/requirements.txt
streamlit
tensorflow
numpy
pillow
rich==13.7.1       # ✅ compatible version
markdown-it-py==3.0.0
mdurl==0.1.2
pygments==2.19.2
